+++ numerical_probabilistic.cpp
@@ -6,7 +6,7 @@
 #include <QtMath>
 #include <QInputDialog>
 using namespace std;
-extern double n;
+extern long long n;
 
 numerical_probabilistic::numerical_probabilistic(QWidget *parent)
     : QMainWindow(parent)
@@ -32,8 +32,8 @@ void numerical_probabilistic::paintEvent(QPaintEvent *event) {
     p.drawText(QRect(30,0,20,100), Qt::AlignCenter, "1");
     p.drawText(QRect(690,347,20,730), Qt::AlignCenter, "1");
     double x, y, z, pi;
-    int count = 0;
-    for (int d = 0; d < n; d++)
+    long long count = 0;
+    for (long long d = 0; d < n; d++)
     {
         x = (float)rand() / RAND_MAX;
         y = (float)rand() / RAND_MAX;
